Added welcome, submission, and wait screen htmls
Added welcome_screen.html, which just tells you which
game you've joined and asks you to wait for the host
to start the game. I'm not sure if this html file
should also be used for the host's welcome screen
when they create a game or if we should make a new one
for that. submit_text.html is where the player can see
the clue from the previous player (if it exists) and
then submit their text. please_wait.html just asks
the player to wait and hit refresh when the player
before them is done.